Force result field if there is no error
Since presence of the result and error field in a response is mutually
exclusive, these fields are tagged omitempty on the struct definition.
This is problematic in the case where an rpc call only returns an error,
and the error is nil. In this case, neither response nor error is added
on the response, which is a violation of the spec.

omitempty can't be removed, as that would cause issues where response is
set even if there is an error, which is also invalid. Therefore, some
handling is added which checks the specific case where there is no error
and no response returned by the function, as this matches the described
situation.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>